+++ app/Http/Resources/EntityLanguageResource.php
@@ -18,6 +18,12 @@ class EntityLanguageResource extends JsonResource
             'language' => $this->when($this->language_id, new LanguageResource($this->whenLoaded('language'))),
             'is_choice' => $this->is_choice,
             'quantity' => $this->when($this->is_choice, $this->quantity),
+            'choice_group' => $this->when($this->choice_group, $this->choice_group),
+            'choice_option' => $this->when($this->choice_option, $this->choice_option),
+            'condition' => $this->when($this->condition_type, fn () => [
+                'type' => $this->condition_type,
+                'language' => new LanguageResource($this->whenLoaded('conditionLanguage')),
+            ]),
         ];
     }
 }